+++ lib/chewy/runtime.rb
@@ -1,9 +1,8 @@
-require 'chewy/runtime/version'
-
 module Chewy
   module Runtime
     def self.version
-      Thread.current[:chewy_runtime_version] ||= Version.new(Chewy.client.info['version']['number'])
+      ActiveSupport::Deprecation.warn "Method 'Chewy::Runtime.version' is deprecated, use 'Chewy.client(name).version' instead."
+      Thread.current[:chewy_runtime_version] ||= Chewy.default_client.version
     end
   end
 end